// Frontend/HospitalManagement/src/app/model/doctor.ts
export class Doctor{
      id?: number;           
  doctorCode: number | undefined;
  doctorName: string | undefined;
  specialize: string | undefined;
  contact: string | undefined;
  availability: string | undefined;  
  email: string | undefined;
  roomNumber: string | undefined;
  description: string | undefined;
  image: string | undefined;
 
}